Deleted _StageBasedExperiment. `trace_model_from_checkpoint` moved to `core.utils`. `get_native_batch` function splitted to two. Docs fixed.
