+++ catalyst/dl/experiment/core.py
@@ -5,7 +5,7 @@ import warnings
 from torch import nn
 from torch.utils.data import DataLoader, Dataset
 
-from catalyst.core import _StageBasedExperiment
+from catalyst.core import _Experiment
 from catalyst.dl import (
     Callback,
     CheckpointCallback,
@@ -23,7 +23,7 @@ from catalyst.utils.tools.settings import STAGE_INFER_PREFIX
 from catalyst.utils.tools.typing import Criterion, Model, Optimizer, Scheduler
 
 
-class Experiment(_StageBasedExperiment):
+class Experiment(_Experiment):
     """
     Super-simple one-staged experiment,
     you can use to declare experiment in code.
