+++ catalyst/dl/utils/__init__.py
@@ -7,7 +7,6 @@ from catalyst.utils import *
 from .torch import get_loader
 from .trace import (
     trace_model,
-    trace_model_from_checkpoint,
     trace_model_from_state,
     get_trace_name,
     save_traced_model,
