+++ catalyst/dl/utils/trace.py
@@ -5,19 +5,15 @@ from pathlib import Path
 from torch import nn
 from torch.jit import ScriptModule, trace, load, save
 
-from catalyst.dl import Experiment, State
+from catalyst.dl import State
 from catalyst.utils.tools.typing import Device, Model
 from catalyst.utils import (
-    import_experiment_and_runner,
     assert_fp16_available,
     is_wrapped_with_ddp,
     set_requires_grad,
     get_requires_grad,
-    unpack_checkpoint,
-    get_native_batch,
+    get_native_batch_from_loaders,
     get_fn_argsnames,
-    load_checkpoint,
-    load_config,
     any2device,
 )
 
@@ -143,89 +139,6 @@ def trace_model(
     return tracer.tracing_result
 
 
-def trace_model_from_checkpoint(
-    logdir: Path,
-    method_name: str,
-    checkpoint_name: str,
-    stage: str = None,
-    loader: Union[str, int] = None,
-    mode: str = "eval",
-    requires_grad: bool = False,
-    opt_level: str = None,
-    device: Device = "cpu",
-):
-    """
-    Traces model using created experiment and runner.
-
-    Args:
-        logdir (Union[str, Path]): Path to Catalyst logdir with model
-        checkpoint_name (str): Name of model checkpoint to use
-        stage (str): experiment's stage name
-        loader (Union[str, int]): experiment's loader name or its index
-        method_name (str): Model's method name that will be
-            used as entrypoint during tracing
-        mode (str): Mode for model to trace (``train`` or ``eval``)
-        requires_grad (bool): Flag to use grads
-        opt_level (str): AMP FP16 init level
-        device (str): Torch device
-
-    Returns:
-        the traced model
-    """
-    config_path = logdir / "configs" / "_config.json"
-    checkpoint_path = logdir / "checkpoints" / f"{checkpoint_name}.pth"
-    print("Load config")
-    config: Dict[str, dict] = load_config(config_path)
-    runner_params = config.get("runner_params", {}) or {}
-
-    # Get expdir name
-    config_expdir = Path(config["args"]["expdir"])
-    # We will use copy of expdir from logs for reproducibility
-    expdir = Path(logdir) / "code" / config_expdir.name
-
-    print("Import experiment and runner from logdir")
-    ExperimentType, RunnerType = import_experiment_and_runner(expdir)
-    experiment: Experiment = ExperimentType(config)
-
-    print(f"Load model state from checkpoints/{checkpoint_name}.pth")
-    if stage is None:
-        stage = list(experiment.stages)[0]
-
-    model = experiment.get_model(stage)
-    checkpoint = load_checkpoint(checkpoint_path)
-    unpack_checkpoint(checkpoint, model=model)
-
-    runner: RunnerType = RunnerType(**runner_params)
-    runner.model, runner.device = model, device
-
-    if loader is None:
-        loader = 0
-    batch = get_native_batch(experiment.get_loaders(stage), loader)
-
-    # function to run prediction on batch
-    def predict_fn(model, inputs, **kwargs):
-        _model = runner.model
-        runner.model = model
-        result = runner.predict_batch(inputs, **kwargs)
-        runner.model = _model
-        return result
-
-    print("Tracing")
-    traced_model = trace_model(
-        model=model,
-        predict_fn=predict_fn,
-        batch=batch,
-        method_name=method_name,
-        mode=mode,
-        requires_grad=requires_grad,
-        opt_level=opt_level,
-        device=device,
-    )
-
-    print("Done")
-    return traced_model
-
-
 def trace_model_from_state(
     state: State,
     method_name: str = "forward",
@@ -263,7 +176,7 @@ def trace_model_from_state(
     # getting batch from loaders and filter it by method argnames
     if loader is None:
         loader = 0
-    batch = get_native_batch(state.loaders, loader)
+    batch = get_native_batch_from_loaders(state.loaders, loader)
     batch = any2device({name: batch[name] for name in method_argnames}, device)
 
     # Dumping previous state of the model, we will need it to restore
@@ -413,7 +326,6 @@ def load_traced_model(
 
 __all__ = [
     "trace_model",
-    "trace_model_from_checkpoint",
     "trace_model_from_state",
     "get_trace_name",
     "save_traced_model",
