+++ catalyst/utils/loader.py
@@ -3,10 +3,29 @@ from typing import Dict, Union  # isort:skip
 from torch.utils.data import DataLoader
 
 
-def get_native_batch(
+def get_native_batch_from_loader(
+        loader: DataLoader,
+        batch_index: int = 0
+):
+    """
+    Returns a batch from experiment loader
+
+    Args:
+        loader (DataLoader): Loader to get batch from
+        batch_index (int): Index of batch to take from dataset of the loader
+    """
+    dataset = loader.dataset
+    collate_fn = loader.collate_fn
+
+    sample = collate_fn([dataset[batch_index]])
+
+    return sample
+
+
+def get_native_batch_from_loaders(
         loaders: Dict[str, DataLoader],
         loader: Union[str, int] = 0,
-        data_index: int = 0
+        batch_index: int = 0
 ):
     """
     Returns a batch from experiment loader
@@ -14,7 +33,7 @@ def get_native_batch(
     Args:
         loaders (Dict[str, DataLoader]): Loaders list to get loader from
         loader (Union[str, int]): Loader name or its index, default is zero
-        data_index (int): Index of batch to take from dataset of the loader
+        batch_index (int): Index of batch to take from dataset of the loader
     """
     if isinstance(loader, str):
         _loader = loaders[loader]
@@ -23,12 +42,13 @@ def get_native_batch(
     else:
         raise TypeError("Loader parameter must be a string or an integer")
 
-    dataset = _loader.dataset
-    collate_fn = _loader.collate_fn
-
-    sample = collate_fn([dataset[data_index]])
-
-    return sample
+    return get_native_batch_from_loader(
+        loader=_loader,
+        batch_index=batch_index
+    )
 
 
-__all__ = ["get_native_batch"]
+__all__ = [
+    "get_native_batch_from_loader",
+    "get_native_batch_from_loaders",
+]
